backend/requirements.txt
Django>=5.1,<6
djangorestframework>=3.15
django-cors-headers>=4.4
Pillow>=10.0
psycopg[binary]>=3.2
python-dotenv>=1.0
